requirements.txt
# 환경 변수 관리
python-dotenv>=1.0.1

# HTTP 요청
requests>=2.32.3

# 데이터 처리 및 과학적 통계 계산
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.0

# 웹 서버 및 데이터베이스
Flask>=3.0.0
PyMySQL>=1.1.0
Flask-SQLAlchemy>=3.1.1

# 한국어 분석 라이브러리
kiwipiepy>=0.17.0

# Open AI LLM
openai>=1.50.0